finalize autodetection
also add mainfest for win10 compat

+++ EnvyUpdate/MainWindow.xaml.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Globalization;
 using System.IO;
 using System.Net;
 using System.Text.RegularExpressions;
@@ -21,7 +22,7 @@ namespace EnvyUpdate
         private readonly string exeloc = System.Reflection.Assembly.GetEntryAssembly().Location;
         private readonly string exepath = Path.GetDirectoryName(System.Reflection.Assembly.GetEntryAssembly().Location) + "\\";
         private readonly string startmenu = Environment.GetFolderPath(Environment.SpecialFolder.StartMenu);
-        private readonly string version = "1.4";
+        private readonly string version = "2.0";
         private string argument = null;
 
         public MainWindow()
@@ -47,11 +48,7 @@ namespace EnvyUpdate
                 MessageBox.Show("Application is already running.");
                 Environment.Exit(1);
             }
-            if (!Directory.Exists(appdata))
-            {
-                Directory.CreateDirectory(appdata);
-            }
-            // Check if application is in Autorun mode and update
+            // Check if application is installed and update
             if (exepath == appdata)
             {
                 try
@@ -65,6 +62,10 @@ namespace EnvyUpdate
                 {
                     // Silently fail.
                 }
+                // Also set correct ticks.
+                chkInstall.IsChecked = true;
+                if (File.Exists(startup + "\\EnvyUpdate.lnk"))
+                    chkAutostart.IsChecked = true;
             }
             if (Util.GetLocDriv() != null)
             {
@@ -85,19 +86,11 @@ namespace EnvyUpdate
                 }
             }
             
-            if (File.Exists(appdata + "nvidia-update.txt"))
-            {
-                chkPortable.IsChecked = false;
-                DispatcherTimer Dt = new DispatcherTimer();
-                Dt.Tick += new EventHandler(Dt_Tick);
-                Dt.Interval = new TimeSpan(5, 0, 0);
-                Dt.Start();
-                Load();
-            }
-            if (File.Exists(startup + "\\EnvyUpdate.lnk"))
-            {
-                chkAutostart.IsChecked = true;
-            }
+            DispatcherTimer Dt = new DispatcherTimer();
+            Dt.Tick += new EventHandler(Dt_Tick);
+            Dt.Interval = new TimeSpan(5, 0, 0);
+            Dt.Start();
+            Load();
         }
 
         private void Dt_Tick(object sender, EventArgs e)
@@ -112,38 +105,17 @@ namespace EnvyUpdate
             //System.Diagnostics.Process.Start("https://github.com/fyr77/EnvyUpdate/");
         }
 
-        private void Grid_Drop(object sender, DragEventArgs e)
-        {
-            if (e.Data.GetDataPresent(DataFormats.FileDrop))
-            {
-                try
-                {
-                    string[] files = (string[])e.Data.GetData(DataFormats.FileDrop);
-                    Load(files);
-                }
-                catch (WebException)
-                {
-                    MessageBox.Show("Network Error. Are you connected to the internet?", "Network Error");
-                }
-            }
-        }
-
         private void Load()
         {
-            FileInfo f = new FileInfo(appdata + "nvidia-update.txt");
-
             int psid;
             int pfid;
             int osid;
-            int langid;
+            //int langid;
 
-            chkPortable.Visibility = Visibility.Hidden;
-            labelDrag.Content = "Drag nvidia.com-cookies.txt here if you have changed your graphics card.";
-            psid = Util.GetData(f.FullName, "ProductSeries");
-            pfid = Util.GetData(f.FullName, "ProductType");
-            osid = Util.GetData(f.FullName, "OperatingSystem");
-            langid = Util.GetData(f.FullName, "Language");
-            gpuURL = "http://www.nvidia.com/Download/processDriver.aspx?psid=" + psid.ToString() + "&pfid=" + pfid.ToString() + "&osid=" + osid.ToString() + "&lid=" + langid.ToString();
+            psid = Util.GetIDs("psid");
+            pfid = Util.GetIDs("pfid");
+            osid = Util.GetIDs("osid");
+            gpuURL = "http://www.nvidia.com/Download/processDriver.aspx?psid=" + psid.ToString() + "&pfid=" + pfid.ToString() + "&osid=" + osid.ToString(); // + "&lid=" + langid.ToString();
             WebClient c = new WebClient();
             gpuURL = c.DownloadString(gpuURL);
             string pContent = c.DownloadString(gpuURL);
@@ -170,50 +142,6 @@ namespace EnvyUpdate
             }
         }
 
-        private void Load(string[] files)
-        {
-            FileInfo f = new FileInfo(files[0]);
-
-            int psid;
-            int pfid;
-            int osid;
-            int langid;
-
-            if (chkPortable.IsChecked == false)
-            {
-                File.Copy(f.FullName, appdata + "nvidia-update.txt", true);
-                f = new FileInfo(appdata + "nvidia-update.txt");
-
-                chkPortable.Visibility = Visibility.Hidden;
-                labelDrag.Content = "Drag nvidia.com-cookies.txt here if you have changed your graphics card.";
-            }
-            psid = Util.GetData(f.FullName, "ProductSeries");
-            pfid = Util.GetData(f.FullName, "ProductType");
-            osid = Util.GetData(f.FullName, "OperatingSystem");
-            langid = Util.GetData(f.FullName, "Language");
-            gpuURL = "http://www.nvidia.com/Download/processDriver.aspx?psid=" + psid.ToString() + "&pfid=" + pfid.ToString() + "&rpf=1&osid=" + osid.ToString() + "&lid=" + langid.ToString() + "&ctk=0";
-            string pContent = null;
-            using (WebClient c = new WebClient())
-            {
-                gpuURL = c.DownloadString(gpuURL);
-                pContent = c.DownloadString(gpuURL);
-            }
-            var pattern = @"\d{3}\.\d{2}&nbsp";
-            Regex rgx = new Regex(pattern);
-            var matches = rgx.Matches(pContent);
-            onlineDriv = Convert.ToString(matches[0]);
-            onlineDriv = onlineDriv.Remove(onlineDriv.Length - 5);
-            textblockOnline.Text = onlineDriv;
-
-            if (localDriv != onlineDriv)
-            {
-                textblockOnline.Foreground = Brushes.Red;
-                buttonDL.Visibility = Visibility.Visible;
-            }
-            else
-                textblockOnline.Foreground = Brushes.Green;
-        }
-
         private void buttonDL_Click(object sender, RoutedEventArgs e)
         {
             System.Diagnostics.Process.Start(gpuURL);
@@ -232,42 +160,46 @@ namespace EnvyUpdate
             }
         }
 
-        private void chkPortable_Unchecked(object sender, RoutedEventArgs e)
+        private void chkInstall_Checked(object sender, RoutedEventArgs e)
         {
             if (chkAutostart != null)
             {
                 chkAutostart.IsEnabled = true;
             }
+            if (exepath != appdata)
+            {
+                if (!Directory.Exists(appdata))
+                {
+                    Directory.CreateDirectory(appdata);
+                }
+                File.Copy(exeloc, appdata + "EnvyUpdate.exe", true);
+                Util.CreateShortcut("EnvyUpdate", startmenu, appdata + "EnvyUpdate.exe", "Nvidia Updater Application.");
+            }
         }
 
-        private void chkPortable_Checked(object sender, RoutedEventArgs e)
+        private void chkInstall_Unchecked(object sender, RoutedEventArgs e)
         {
             if (chkAutostart != null)
             {
                 chkAutostart.IsEnabled = false;
                 chkAutostart.IsChecked = false;
             }
+            if (Directory.Exists(appdata))
+            {
+                File.Delete(appdata + "EnvyUpdate.exe");
+                File.Delete(startup + "\\EnvyUpdate.lnk");
+                File.Delete(startmenu + "\\EnvyUpdate.lnk");
+            }
         }
 
         private void chkAutostart_Checked(object sender, RoutedEventArgs e)
         {
-            if (exepath != appdata)
-            {
-                File.Copy(exeloc, appdata + "EnvyUpdate.exe", true);
-                Util.CreateShortcut("EnvyUpdate", startup, appdata + "EnvyUpdate.exe", "Nvidia Updater Application.");
-                Util.CreateShortcut("EnvyUpdate", startmenu, appdata + "EnvyUpdate.exe", "Nvidia Updater Application.");
-            }
-            else
-            {
-                chkAutostart.IsEnabled = false;
-            }
+            Util.CreateShortcut("EnvyUpdate", startup, appdata + "EnvyUpdate.exe", "Nvidia Updater Application.");
         }
 
         private void chkAutostart_Unchecked(object sender, RoutedEventArgs e)
         {
-            File.Delete(appdata + "EnvyUpdate.exe");
             File.Delete(startup + "\\EnvyUpdate.lnk");
-            File.Delete(startmenu + "\\EnvyUpdate.lnk");
         }
 
         private void Window_Closing(object sender, System.ComponentModel.CancelEventArgs e)
